+++ LMeter/Config/VisibilityConfig.cs
@@ -24,7 +24,9 @@ namespace LMeter.Config
         public bool HideOutsideCombat = false;
         public bool HideOutsideDuty = false;
         public bool HideWhilePerforming = false;
+        public bool HideInPvP = false;
         public bool HideInGoldenSaucer = false;
+        public bool HideInFieldOperations = false;
         public bool HideIfNotConnected = false;
         public bool ShouldClip = true;
 
@@ -59,7 +61,17 @@ namespace LMeter.Config
                 return false;
             }
 
+            if (this.HideInPvP && CharacterState.IsInPvP())
+            {
+                return false;
+            }
+
             if (this.HideInGoldenSaucer && CharacterState.IsInGoldenSaucer())
+            {
+                return false;
+            }
+
+            if (this.HideInFieldOperations && CharacterState.IsInFieldOperation())
             {
                 return false;
             }
@@ -81,7 +93,9 @@ namespace LMeter.Config
                 ImGui.Checkbox("Hide Outside Combat", ref this.HideOutsideCombat);
                 ImGui.Checkbox("Hide Outside Duty", ref this.HideOutsideDuty);
                 ImGui.Checkbox("Hide While Performing", ref this.HideWhilePerforming);
+                ImGui.Checkbox("Hide In PvP", ref this.HideInPvP);
                 ImGui.Checkbox("Hide In Golden Saucer", ref this.HideInGoldenSaucer);
+                ImGui.Checkbox("Hide In Field Operations", ref this.HideInFieldOperations);
                 ImGui.Checkbox("Hide While Not Connected to ACT", ref this.HideIfNotConnected);
                 ImGui.Checkbox("Hide When Covered by Game UI Window", ref this.ShouldClip);